+++ fuzz_test.go
@@ -40,7 +40,10 @@ func FuzzLunarPhase(f *testing.F) {
 		if date.Year() < 1800 || date.Year() > 2200 {
 			return
 		}
-		p := LunarPhase(date)
+		p, err := LunarPhase(date)
+		if err != nil {
+			return
+		}
 		if math.IsNaN(p.Illumination) {
 			t.Error("NaN illumination")
 		}